requirements.txt
dspy
python-dotenv
fpdf